requirements.txt
flask==2.0.1
ccxt==3.0.0
pandas==1.4.2
numpy==1.22.3
ta-lib==0.4.24
alpaca-trade-api==2.3.0
python-dotenv==0.19.2
gunicorn==20.1.0
yfinance==0.2.12
pybit==2.4.1 